`Shamrock`: fix `group_id` | `user_id`

+++ xposed/src/main/java/moe/fuqiuluo/shamrock/remote/action/handlers/GetStatus.kt
@@ -14,7 +14,7 @@ internal object GetStatus: IActionHandler() {
         val curUin = runtime.currentAccountUin
         return resultToString(true, Status.Ok, listOf(
             BotStatus(
-                Self("qq", curUin), runtime.isLogin, status = "正常", good = runtime.isLogin
+                Self("qq", curUin.toLong()), runtime.isLogin, status = "正常", good = runtime.isLogin
             )
         ), echo = session.echo)
     }
